Add Pseudolocalizer#pseudolocalize as a standalone lib
Having access to a lib will decouple the feature from
the I18n backend, allowing us to call pseudolocalize
from outside the context of an I18n backend.
