+++ lib/pseudolocalization/version.rb
@@ -1,3 +1,3 @@
 module Pseudolocalization
-  VERSION = "0.5.0"
+  VERSION = "0.6.0"
 end
